[ADD] add new file sliced_result
