+++ source/app/blueprints/graphql/sliced_result.py
@@ -1,0 +1,32 @@
+#  IRIS Source Code
+#  Copyright (C) 2024 - DFIR-IRIS
+#  contact@dfir-iris.org
+#
+#  This program is free software; you can redistribute it and/or
+#  modify it under the terms of the GNU Lesser General Public
+#  License as published by the Free Software Foundation; either
+#  version 3 of the License, or (at your option) any later version.
+#
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+#  Lesser General Public License for more details.
+#
+#  You should have received a copy of the GNU Lesser General Public License
+#  along with this program; if not, write to the Free Software Foundation,
+#  Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+
+class SlicedResult:
+    def __init__(self, data, start_offset, total):
+        self._data = data
+        self._start = start_offset
+        self._total = total
+
+    def __getitem__(self, index: slice) -> any:
+        x = (index.start - self._start)
+        y = (index.stop - self._start)
+        return self._data[x: y]
+
+    def __len__(self) -> int:
+        return self._total
+
