+++ tests/tests.py
@@ -880,7 +880,6 @@ class Tests(TestCase):
         }
         response = self._subject.execute_graphql_query(payload)
         severity_id = response['data']['caseCreate']['case']['severityId']
-        print(severity_id)
         payload = {
             'query': f'''query {{ cases(severityId: {severity_id}) 
                     {{ edges {{ node {{ severityId }} }} }} }}'''
